Fix type conversion warnings in PBRefController, PBQLOutlineView, and PBGitIndex
- PBRefController.m:286: Use numberWithInteger for NSInteger row parameter
- PBQLOutlineView.m:70: Change int row to NSInteger row to match rowAtPoint return type
- PBGitIndex.m:285,289: Change int filesCount and loopTo to NSUInteger to match array count type

+++ Classes/git/PBGitIndex.m
@@ -282,11 +282,11 @@ NSString *PBGitIndexOperationFailed = @"PBGitIndexOperationFailed";
 {
 	// Do staging files by chunks of 1000 files each, to prevent program freeze (because NSPipe has limited capacity)
 	
-	int filesCount = [stageFiles count];
+	NSUInteger filesCount = [stageFiles count];
 	
 	// Prepare first iteration
 	int loopFrom = 0;
-	int loopTo = 1000;
+	NSUInteger loopTo = 1000;
 	if (loopTo > filesCount)
 		loopTo = filesCount;
 	int loopCount = 0;
